Implement universal backup manager
